+++ ttttt/src/math/normalize_angle.c
@@ -6,13 +6,13 @@
 /*   By: nwakour <nwakour@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/20 16:23:16 by nwakour           #+#    #+#             */
-/*   Updated: 2020/12/25 18:11:18 by nwakour          ###   ########.fr       */
+/*   Updated: 2020/12/28 17:09:04 by nwakour          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../cub3d.h"
 
-float	normalize_angle(float angle)
+double	normalize_angle(double angle)
 {
 	angle = remainder(angle, (2 * M_PI));
 	if (angle < 0)
